# api/openapi.yaml
openapi: 3.0.1
info:
  title: VMClarity APIs
  version: 1.0.0
paths:
...
  /assetScanEstimations:
    get:
      summary: Get asset scan estimations according to the given filters
      operationId: GetAssetScanEstimations
      parameters:
        - $ref: '#/components/parameters/odataFilter'
        - $ref: '#/components/parameters/odataSelect'
        - $ref: '#/components/parameters/odataCount'
        - $ref: '#/components/parameters/odataTop'
        - $ref: '#/components/parameters/odataSkip'
        - $ref: '#/components/parameters/odataExpand'
        - $ref: '#/components/parameters/odataOrderBy'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetScanEstimations'

        default:
          $ref: '#/components/responses/UnknownError'
    post:
      summary: Create an asset scan estimation for a specified asset
      operationId: PostAssetScanEstimations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetScanEstimation'
        required: true
      responses:
        201:
          description: AssetScanEstimation was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetScanEstimation'
        400:
          description: Invalid asset scan estimation supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        409:
          description: Conflicting asset scan estimation found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetScanEstimationExists'
        default:
          $ref: '#/components/responses/UnknownError'
      x-codegen-request-body-name: body
  /assetScanEstimations/{assetScanEstimationID}:
    get:
      summary: Get an asset scan stimation.
      operationId: GetAssetScanEstimationsAssetScanEstimationID
      parameters:
        - $ref: '#/components/parameters/assetScanEstimationID'
        - $ref: '#/components/parameters/odataSelect'
        - $ref: '#/components/parameters/odataExpand'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetScanEstimation'
        404:
          description: AssetScanEstimation ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          $ref: '#/components/responses/UnknownError'
    put:
      summary: Update an asset scan estimation.
      operationId: PutAssetScanEstimationsAssetScanEstimationID
      parameters:
        - $ref: '#/components/parameters/assetScanEstimationID'
        - $ref: '#/components/parameters/ifmatch'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetScanEstimation'
        required: true
      responses:
        200:
          description: Updated an asset scan estimation successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetScanEstimation'
        400:
          description: Invalid asset scan estimation supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        404:
          description: AssetScanEstimation ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        409:
          description: Conflicting asset scan estimation found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetScanEstimationExists'
        412:
          description: Etag didn't match.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          $ref: '#/components/responses/UnknownError'
      x-codegen-request-body-name: body
    patch:
      summary: Patch an asset scan estimation
      operationId: PatchAssetScanEstimationsAssetScanEstimationID
      parameters:
        - $ref: '#/components/parameters/assetScanEstimationID'
        - $ref: '#/components/parameters/ifmatch'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetScanEstimation'
        required: true
      responses:
        200:
          description: Updated an asset scan estimation successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetScanEstimation'
        400:
          description: Invalid asset scan estimation supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        404:
          description: AssetScanEstimation ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        409:
          description: Conflicting asset scan estimation found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetScanEstimationExists'
        412:
          description: Etag didn't match.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          $ref: '#/components/responses/UnknownError'
      x-codegen-request-body-name: body
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        message:
          type: string
          readOnly: true
      description: An object that is returned in all cases of failures.
...
    CloudProvider:
      type: string
      enum:
        - AWS
        - Azure
        - GCP
        - Docker
        - External
...
    ScanFindingsSummary:
      description: A summary of the scan findings.
      type: object
      properties:
        totalPackages:
          type: integer
        totalExploits:
          type: integer
        totalMalware:
          type: integer
        totalMisconfigurations:
          type: integer
        totalRootkits:
          type: integer
        totalSecrets:
          type: integer
        totalVulnerabilities:
          $ref: '#/components/schemas/VulnerabilityScanSummary'
    VulnerabilityScanSummary:
      description: A summary of number of vulnerabilities found per severity.
      type: object
      properties:
        totalCriticalVulnerabilities:
          type: integer
        totalHighVulnerabilities:
          type: integer
        totalMediumVulnerabilities:
          type: integer
        totalLowVulnerabilities:
          type: integer
        totalNegligibleVulnerabilities:
          type: integer
...
    ScanFamiliesConfig:
      type: object
      description: The configuration of the scanner families within a scan config
      properties:
        sbom:
          $ref: '#/components/schemas/SBOMConfig'
        vulnerabilities:
          $ref: '#/components/schemas/VulnerabilitiesConfig'
        malware:
          $ref: '#/components/schemas/MalwareConfig'
        rootkits:
          $ref: '#/components/schemas/RootkitsConfig'
        secrets:
          $ref: '#/components/schemas/SecretsConfig'
        misconfigurations:
          $ref: '#/components/schemas/MisconfigurationsConfig'
        exploits:
          $ref: '#/components/schemas/ExploitsConfig'
    VulnerabilitiesConfig:
      type: object
      properties:
        enabled:
          type: boolean
        scanners:
          type: array
          items:
            type: string
    SBOMConfig:
      type: object
      properties:
        enabled:
          type: boolean
        analyzers:
          type: array
          items:
            type: string
    MalwareConfig:
      type: object
      properties:
        enabled:
          type: boolean
        scanners:
          type: array
          items:
            type: string
    RootkitsConfig:
      type: object
      properties:
        enabled:
          type: boolean
        scanners:
          type: array
          items:
            type: string
    SecretsConfig:
      type: object
      properties:
        enabled:
          type: boolean
        scanners:
          type: array
          items:
            type: string
    MisconfigurationsConfig:
      type: object
      properties:
        enabled:
          type: boolean
        scanners:
          type: array
          items:
            type: string
    ExploitsConfig:
      type: object
      properties:
        enabled:
          type: boolean
        scanners:
          type: array
          items:
            type: string
...
    ScannerInstanceCreationConfig:
      type: object
      description: Configuration of scanner instance
      properties:
        useSpotInstances:
          type: boolean
        retryMaxAttempts:
          type: integer
        maxPrice:
          type: string
      required:
        - useSpotInstances
...
    SecurityGroup:
      type: object
      description: general cloud security group
      properties:
        id:
          type: string
          minLength: 1
      required:
        - id
      additionalProperties: false
    Tag:
      type: object
      description: general cloud tag / label
      properties:
        key:
          type: string
        value:
          type: string
      required:
        - key
        - value
      additionalProperties: false
    RootVolume:
      type: object
      description: Information about VM root volume
      properties:
        sizeGB:
          type: integer
        encrypted:
          type: string
          enum:
            - Yes
            - No
            - Unknown
      required:
        - sizeGB
        - encrypted
      additionalProperties: false
...
    AssetRelationship:
      type: object
      description: Describes a relationship to an asset which can be expanded.
      properties:
        id:
          type: string
        revision:
          type: integer
          readOnly: true
        assetInfo:
          $ref: '#/components/schemas/AssetType'
          readOnly: true
        scansCount:
          description: Total number of scans that have ever run for this asset
          type: integer
          readOnly: true
        firstSeen:
          type: string
          format: date-time
        terminatedOn:
          type: string
          format: date-time
          readOnly: true
        lastSeen:
          type: string
          format: date-time
          readOnly: true
        summary:
          $ref: '#/components/schemas/ScanFindingsSummary'
          readOnly: true
      required: ['id']
...
    AssetType:
      type: object
      anyOf:
        - $ref: '#/components/schemas/VMInfo'
        - $ref: '#/components/schemas/PodInfo'
        - $ref: '#/components/schemas/DirInfo'
        - $ref: '#/components/schemas/ContainerImageInfo'
        - $ref: '#/components/schemas/ContainerInfo'
      discriminator:
        propertyName: objectType
        mapping:
          VMInfo: '#/components/schemas/VMInfo'
          PodInfo: '#/components/schemas/PodInfo'
          DirInfo: '#/components/schemas/DirInfo'
          ContainerImageInfo: '#/components/schemas/ContainerImageInfo'
          ContainerInfo: '#/components/schemas/ContainerInfo'
    VMInfo:
      type: object
      properties:
        objectType:
          type: string
        instanceID:
          type: string
        instanceProvider:
          $ref: '#/components/schemas/CloudProvider'
        location:
          type: string
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
          nullable: true
        securityGroups:
          type: array
          items:
            $ref: '#/components/schemas/SecurityGroup'
          nullable: true
        image:
          type: string
        instanceType:
          type: string
        platform:
          type: string
        launchTime:
          type: string
          format: date-time
        rootVolume:
          $ref: '#/components/schemas/RootVolume'
      required:
        - objectType
        - instanceID
        - location
        - instanceType
        - image
        - platform
        - launchTime
        - rootVolume
    PodInfo:
      type: object
      properties:
        objectType:
          type: string
        podName:
          type: string
        location:
          type: string
      required:
        - objectType
    ContainerInfo:
      type: object
      properties:
        objectType:
          type: string
        containerName:
          type: string
        # TODO(paralta) Check if image needs to be a required property
        image:
          $ref: '#/components/schemas/ContainerImageInfo' 
        id:
          type: string
        createdAt:
          type: string
          format: date-time
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
          nullable: true
      required:
        - objectType
    ContainerImageInfo:
      type: object
      properties:
        objectType:
          type: string
        id:
          type: string
        name:
          type: string
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
          nullable: true
        architecture:
          type: string
        os:
          type: string
        size:
          type: integer
      required:
        - objectType 
    DirInfo:
      type: object
      properties:
        objectType:
          type: string
        dirName:
          type: string
        location:
          type: string
      required:
        - objectType
...
    AssetScanTemplate:
      type: object
      x-go-type-name: "AssetScanTemplate"
      properties:
        scanFamiliesConfig:
          $ref: '#/components/schemas/ScanFamiliesConfig'
        scannerInstanceCreationConfig:
          $ref: '#/components/schemas/ScannerInstanceCreationConfig'
...
    AssetScanEstimationState:
      type: object
      properties:
        state:
          type: string
          enum:
            - Pending
            - Aborted
            - Failed
            - Done
        lastTransitionTime:
          type: string
          format: date-time
        errors:
          type: array
          items:
            type: string
          nullable: true
...
    Estimation:
      type: object
      properties:
        time:
          description: Total time the scan will take (seconds)
          type: integer
        size:
          description: Total size of the scan (GB)
          type: integer
        cost:
          description: Total cost of the scan ($)
          type: integer
          format: float
    AssetScanEstimations:
      type: object
      properties:
        count:
          description: Total AssetScanEstimations count according to the given filters
          type: integer
        items:
          description: List of AssetScanEstimations according to the given filters
          type: array
          items:
            $ref: '#/components/schemas/AssetScanEstimation'
    ScanEstimationFakeRelationship:
      type: object
      properties:
        id:
          type: string
    AssetScanEstimation:
      type: object
      properties:
        id:
          type: string
        asset:
          $ref: '#/components/schemas/AssetRelationship'
        # TODO (erezf) replace with ScanEstimationRelationship object (?)
        scanEstimation:
          $ref: '#/components/schemas/ScanEstimationFakeRelationship'
        state:
          $ref: '#/components/schemas/AssetScanEstimationState'
        estimation:
          $ref: '#/components/schemas/Estimation'
        assetScanTemplate:
          $ref: '#/components/schemas/AssetScanTemplate'
    AssetScanEstimationExists:
      type: object
      properties:
        message:
          description: Describes which unique constraint combination causes the conflict.
          type: string
          readOnly: true
        assetScanEstimation:
          $ref: '#/components/schemas/AssetScanEstimation'
...
  responses:
...
    UnknownError:
      description: Unknown error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
  parameters:
...
    odataFilter:
      name: "$filter"
      in: query
      schema:
        type: string
    odataSelect:
      name: "$select"
      in: query
      schema:
        type: string
    odataCount:
      name: "$count"
      in: query
      schema:
        type: boolean
    odataTop:
      name: "$top"
      in: query
      schema:
        type: integer
    odataSkip:
      name: "$skip"
      in: query
      schema:
        type: integer
    odataExpand:
      name: "$expand"
      in: query
      schema:
        type: string
    odataOrderBy:
      name: "$orderby"
      in: query
      schema:
        type: string
      x-go-name: "OrderBy"
...
    assetScanEstimationID:
      name: assetScanEstimationID
      in: path
      required: true
      schema:
        type: string
    ifmatch:
      name: If-Match
      in: header
      schema:
        type: integer